Remove d3
