+++ package.json
@@ -22,7 +22,6 @@
   "dependencies": {
     "backbone": "^1.3.3",
     "body-parser": "^1.14.2",
-    "d3": "^4.13.0",
     "express": "^4.13.3",
     "google-trends-api": "^4.4.0",
     "jquery": "^3.3.1",
